feat: added report service + removed unused code
