+++ app/catalog/views.py
@@ -2,12 +2,8 @@ from flask import abort, redirect, render_template, request, session, url_for
 
 from app.catalog import blueprint
 from app.catalog.forms import UploadForm
-from app.catalog.utils import (
-    gen_report,
-    read_uploaded_set_excel_file,
-    send_temp_file,
-)
-from app.services import sets_srv
+from app.catalog.utils import read_uploaded_set_excel_file, send_temp_file
+from app.services import report_srv, sets_srv
 
 
 @blueprint.route("/")
@@ -61,7 +57,9 @@ def report():
             abort(400)
 
         # Generate report
-        set_report = gen_report(parts_df, minifigs_parts_df, elements_df)
+        set_report = report_srv.generate_report(
+            parts_df, minifigs_parts_df, elements_df
+        )
         return render_template(
             "report.html",
             parts=set_report["parts"],
